requirements.txt
dvc
dvc[gdrive]
sklearn
pandas
pyyaml